# Improve performance of functions with dynamic arguments