@osa1 2025-04-16
This change doesn't need to be a breaking change. We can add a deprecation to `translate` though as I would think that it's too slow for what it's intended for.