@spydon 2025-04-16
> This change doesn't need to be a breaking change. We can add a deprecation to `translate` though as I would think that it's too slow for what it's intended for.

Yeah, it wouldn't be a breaking change right now. But if following semver, marking it as deprecated is just the first step towards a breaking change, when removing the deprecated method major should be bumped (https://semver.org/#how-should-i-handle-deprecating-functionality).

I'm not too fond of all the `byType` methods, but I also can't come up with a better solution, and it is definitely better than taking `dynamic` as an argument.