@osa1 2025-04-29
> by all means revert if you don't like the last commit

I dropped your commit as it changed benchmark results and made them worse, and then I had to debug this because I wasn't able to reproduce the benchmark results in my PR description. 

(The "cleanup" change calls less efficient function from more efficient one and makes `translateByVector3` benchmark worse)