@osa1 2025-04-30
@mraleph @rakudrama @mkustermann I think I addressed all of the feedback. Could you PTAL? Start with the PR description, I just applied the same changes to other functions that take `dynamic` arguments.

@kevmoo as far as I know still no one from the backend teams can approve reviews, can we give at least one person from each backend review rights? They already reviewed this PR, they just officially can't approve.

(Same as SDK, it makes sense for changes to libraries to be reviewed by backend teams to catch any performance or binary size issues)