@osa1 2025-04-30
> move stores up as far as possible

Done. I haven't benchmarked this change in isolation.

> have single implementation that all others delegate to

Done. Doesn't affect performance when I inline everything.

> make functions only have a single return & mark as @pragma('dart2js:prefer-inline')

Done. Improves dart2js performance.

> access [15] first

NOT done: doesn't make any difference in dart2wasm and AOT, but makes dart2js perform worse.