@rakudrama 2025-04-30
> Generally LGTM with the following comments
> 
> * move stores up as far as possible

I'm worried about moving stores. There are constructors that take 'storage' as an input, so there can be aliasing via typed data views.
If you move a store, there is an obligation to explain to the reader why the order is safe for aliased inputs.
Yes, there are lots of bugs already with aliasing, but lets not add more.

> * access `[15]` first

Can we make the statement check `_m4storage[15];` work on all platforms? It works for VM-AOT and dart2js.
Moving statements for the effect of moving `[15]` first is less readable and open to changing read-write order that may affect aliased inputs.

> * have single implementation that all others delegate to
> * make functions only have a single return & mark as `@pragma('dart2js:prefer-inline')`
> 
> If the last point improves performance in dart2js similar to dart2wasm & vm, then do we get meaningful perf benefits from having the `{translate,scale,...}By{Double,...}` operations?