@osa1 2025-05-01
> I'm worried about moving stores. There are constructors that take 'storage' as an input, so there can be aliasing via typed data views.
> If you move a store, there is an obligation to explain to the reader why the order is safe for aliased inputs.
> Yes, there are lots of bugs already with aliasing, but lets not add more.

Which function(s) do you mean that can have aliasing bugs?

The code with stores moved up is this:

```dart
void translateByDouble(double tx, double ty, double tz, double tw) {
  final t1 = _m4storage[0] * tx +
      _m4storage[4] * ty +
      _m4storage[8] * tz +
      _m4storage[12] * tw;
  _m4storage[12] = t1;

  final t2 = _m4storage[1] * tx +
      _m4storage[5] * ty +
      _m4storage[9] * tz +
      _m4storage[13] * tw;
  _m4storage[13] = t2;

  final t3 = _m4storage[2] * tx +
      _m4storage[6] * ty +
      _m4storage[10] * tz +
      _m4storage[14] * tw;
  _m4storage[14] = t3;

  final t4 = _m4storage[3] * tx +
      _m4storage[7] * ty +
      _m4storage[11] * tz +
      _m4storage[15] * tw;
  _m4storage[15] = t4;
}
```

There's only one array being used here, so aliasing can't be an issue.

> Can we make the statement check _m4storage[15]; work on all platforms? 

I reported benchmarks with this here: https://github.com/google/vector_math.dart/pull/345#discussion_r2063389469.

It doesn't make any difference in any of: AOT, dart2js, dart2wasm.

It's difficult to make it work with dart2wasm because wasm-opt doesn't track accessed locations and omit bounds checks based on that. As mentioned in the thread, I reported this use case to the wasm-opt team and but it's in the backlog with low priority.