@osa1 2025-05-01
Bumped major version instead as this is a breaking change. (new membrers added to non-final class)