@kevmoo 2025-05-12
https://github.com/dart-lang/sdk/commit/3b444c590a72a675ba5e9660f0c1b0fa512e86d6 landed and I'm guessing this rolled to Google3. 

Should we do a release and get this rolled to Flutter?

Should we do a verification pass on Flutter first?