@osa1 2025-05-13
I think ideally for a package this small and simple we shouldn't need rolls for testing, we should be able to test it properly with just `dart test` and release.

I don't know how comprehensive testing in this package is as I did just one commit to the package, but I see tests for all the changed functions: `translate`, `leftTranslate`, `scale`, `multiply`.

Given that the changes are simple, and we have unit tests covering each of the changed functions, I would just publish.